+++ tests/test_app/resources/store.py
@@ -4,7 +4,7 @@ from marshmallow import fields
 import tests.test_app.models as m
 from awokado import custom_fields
 from awokado.consts import CREATE, READ, BULK_CREATE
-from awokado.utils import ReadContext
+from awokado.request import ReadContext
 from tests.test_app.resources.base import Resource